Moved commit test check into common. Added to new_sequence test
